requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.1
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.13.0
pydantic_core==2.46.0
python-dotenv==1.2.1
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
